feat(maintenance): restart script generates nginx config from template (#248)
What does this PR do / solve?
-----------------------------

The configuration of our `nginx` HTTP proxy contains redundant definitions in `/etc/nginx/sites-available`, in order to allow no-downtime deployments. (which rely on switching from a port to another)

Redundancy makes it harder and riskier to update the `nginx` configuration.

Overview of changes
-------------------

This PR makes the `restart.sh` script re-generate these two `nginx` configuration from a template file. => The template becomes the source of truth, and it's versionned in this repo.

How to test this PR?
--------------------

Test in production :-p