fix(bookmarklet): refactor bookmarklet and add tests for #262 (#281)
Contributes to #262. Follow up of #280.

## What does this PR do / solve?

As mentioned in #262 following the reduction of our YouTube API quota, we need to make several changes to the bookmarklet in order to save some quota while keeping a good user experience. E.g. track titles should be extracted from the page instead of queried from YouTube API.

As these changes are quite structural and may introduce bugs in the bookmarklet, and as we don't have time to test it manually after each change, it's time to refactor the code in order to make that code testable by automated unit tests.

This PR is the second step of that refactor: it enables adaptive metadata enrichment for the page itself (e.g. youtube track page), while preventing duplicates in the results.

## Overview of changes

See the diff, and disable the display of whitespace changes.

- add `detectYouTubePageTrack()`
- group ui calls
- bump to bookmarklet v2.5 (from v2.4)
- add unit tests

## How to test this PR?

```sh
$ nvm use
$ node_modules/.bin/mocha test/unit/bookmarklet-tests.js
$ docker-compose up --build -d
$ npm run docker:seed && node_modules/.bin/wdio wdio.conf.js
``` 

Also, you can test the bookmarklet manually, from your web browser, by following the instructions provided in the unit test file.