fix(ui): never-ending loading animation on empty search (#260)
Closes #226.

## What does this PR do / solve?

The console javascript error appears because there is a results condition in the file app/templates/searchPage.html.

This condition wraps the tab_tracks element that the file public/js/searchPage.js is looking for. I found 2 front end solutions to this specific issue.

## Overview of changes

First we can wrap the searchPage.js file within the same condition as the html and give some sort of html message stating an invalid search query has been inputed.

Or we can have a condition within the searchPage.js file checking for the element. If the element is not found then the page will redirect to the homepage. This is the technique I found being used by google and amazon when I added an empty query in the url. In this pull request I have used this second solution.

## How to test this PR?

I used the javascript debugger and the console to add a javascript break before entering the condition. After a valid and  empty search there were no errors showing and the page redirected successfully when an empty search was given.