fix(error): 404 redirect when clicking load more on account stream (#296)
Closes #295 .

## What does this PR do / solve?

It looks like the user gets redirected to the 404 page because of a javascript error that involves the template condition inside the script tag. This error prevents the function from running and rendering the correct url for the button.

## Overview of changes

I added multiline comments around template conditions inside the script tags.

## How to test this PR?

Non regression tests were added and merged to `master`, in PR #299.