fix(tests): create dummy posts from cypress tests instead of from initdb_testing (#304)
Contributes to #199.

## What does this PR do / solve?

We recently added 21 hard-coded fake posts in `config/initdb_testing.js`, so that the pagination test could run.

These 21 posts are just required by that tests, so they should be added from that test only.

## Overview of changes

Add a `postDummyTracks()` command that generates and inserts a given number of posts for the currently logged-in user. => Call it from pagination tests.