feat(ci): add version as tag, when publishing on Docker Hub (#306)
So far, we've been publishing images to Docker Hub systematically with the `latest` tag. This means that we overwrite the previous version of the image everytime, instead of allowing users to download a specific version of the Openwhyd docker image.

After merging this PR, every release should be listed on Docker Hub, with their version number as tag, in addition to the `latest` tag.