feat: Add alphabetical order to playlists page (#372)
Closes #371 

## What does this PR do / solve?

Add playlist sort feature that includes an alphabetized option

## Overview of changes

Added sort menu to the playlist header with a sort by date option and an alphabetized option

![sort-menu-initial](https://user-images.githubusercontent.com/13477167/97369266-d3dc6080-1869-11eb-99d6-5811ff423640.png)

![sort-menu-hover-alphabetized](https://user-images.githubusercontent.com/13477167/97369289-e35ba980-1869-11eb-991a-f97a3d7b78b5.png)

## How to test this PR?

Use the profile side nav and click on the playlist menu item. On the playlist page you will now have a sort menu next to the playlist main header.