fix(bookmarklet): Remove PlayemJS from bookmarklet (#408)
## What does this PR do / solve?

PlayemJS has been maintained for Openwhyd, to play and detect tracks, by communicating with the APIs of the streaming sources we support.

While PlayemJS has supported more sources than most alternative players we have found online, its codebase has become hard to maintain, so it suffers from a lack of updates, which lead to numerous problems taking a lot of time and effort to fix: e.g. #272, #262, #192, #190, #143, #132, #128, #16...

In order to simplify Openwhyd's codebase and make it easier for us to replace PlayemJS with an alternative, let's start by trying to remove it from the bookmarklet and see if anybody misses it from there.